+++ sdk/storage/azure-storage-blob-nio/CHANGELOG.md
@@ -7,7 +7,7 @@
 ### Breaking Changes
 
 ### Bugs Fixed
-- Throw java.nio.NoSuchFileException instead of raw java.io.IOException for non existent blob file attributes request
+- Reading attributes for a non-existent blob now throws java.nio.file.NoSuchFileException instead of a raw java.io.IOException.
 
 ### Other Changes