+++ sdk/storage/azure-storage-blob-nio/CHANGELOG.md
@@ -1,9 +1,9 @@
 # Release History
 
 ## 12.0.0-beta.37 (Unreleased)
-- Extend azure nio filesystem url to support a unique identifer: `azb://?endpoint=<account_endpoint>&uid=<unique_id>`
 
 ### Features Added
+- Extended Azure NIO filesystem URI support to allow a unique identifier: `azb://?endpoint=<account_endpoint>&uid=<unique_id>`
 
 ### Breaking Changes